+++ src/drivers/host/timer.c
@@ -7,7 +7,7 @@
 //         Rander Wang <rander.wang@intel.com>
 //         Janusz Jankowski <janusz.jankowski@linux.intel.com>
 
-#include "testbench/timer.h"
+#include <sof/drivers/timer.h>
 
 void platform_host_timestamp(struct comp_dev *host,
 			     struct sof_ipc_stream_posn *posn)
@@ -19,3 +19,9 @@ void platform_dai_timestamp(struct comp_dev *dai,
 			    struct sof_ipc_stream_posn *posn)
 {
 }
+
+uint64_t platform_timer_get(struct timer *timer)
+{
+	return 0;
+}
+
