+++ src/platform/library/CMakeLists.txt
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: BSD-3-Clause
+
+add_subdirectory(lib)
+add_subdirectory(schedule)
+
+add_local_sources(sof panic.c)
+
