+++ src/platform/library/lib/CMakeLists.txt
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: BSD-3-Clause
+
+add_local_sources(sof
+	alloc.c
+	clk.c
+	trace.c
+)
