+++ src/platform/library/lib/clk.c
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: BSD-3-Clause
+
+#include <sof/lib/clk.h>
+
+uint64_t clock_ms_to_ticks(int clock, uint64_t ms)
+{
+	return 0;
+}
