+++ src/platform/library/lib/trace.c
@@ -8,8 +8,7 @@
 //         Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
 
 #include <stdint.h>
-#include "testbench/common_test.h"
-#include "testbench/trace.h"
+#include <sof/trace/trace.h>
 
 /* enable trace by default in testbench */
 int test_bench_trace = 1;
@@ -25,19 +24,3 @@ char *get_trace_class(uint32_t trace_class)
 	return "unknown";
 }
 
-/* print debug messages */
-void debug_print(char *message)
-{
-	if (debug)
-		printf("debug: %s", message);
-}
-
-/* enable trace in testbench */
-void tb_enable_trace(bool enable)
-{
-	test_bench_trace = enable;
-	if (enable)
-		debug_print("trace print enabled\n");
-	else
-		debug_print("trace print disabled\n");
-}
