+++ src/platform/library/panic.c
