+++ src/platform/library/schedule/CMakeLists.txt
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: BSD-3-Clause
+
+add_local_sources(sof
+	schedule.c
+	ll_schedule.c
+)
